refactor: 마이페이지 구성 수정
- lazy 적용
- 폴더 구조 변경
